refactor(tests): create helpers module and extract sync_basic.rs
Move shared test helpers (builders, mocks, fixtures) into
tests/helpers/ and extract the basic sync tests (happy path,
skip on digest match, source errors, multi-layer, cross-repo
mount) into sync_basic.rs.

Part of structural split to make the 12K-line monolith navigable.
Zero behavioral change -- same tests, same assertions.
